Update to incremental building of GIF pages
